REMOVES the build directory from the tree..
The reason for this is that the build tree is specific to the cumputer
on which it is compiled, and serves no purpose to keep it tracked by
git. With this commit, the build directory is no longer part of the
tree, and will be ignored by git in the future. running make will work
like normal, but only the final ROM generated will be uploaded when you
"git push", and none of the intermediate files, as all the other files
need to be regenerated.

Signed-off-by: Spartan97 <icelax1992@comcast.net>